changes fixed
